+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- `checkout-summary` component to `Cart`.
+
+### Changed
+
+- Improved cart layout and responsiveness.
+
 ## [0.1.0] - 2019-08-13
 
 ### Added